services/api/pyproject.toml
[project]
name = "applylens-api"
version = "0.1.0"
description = "ApplyLens FastAPI backend"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "gunicorn",
  "pydantic-settings",
  "SQLAlchemy>=2.0",
  "psycopg2-binary",
  "alembic",
  "python-multipart",
  "orjson",
  "elasticsearch>=8.12,<9",
  "httpx",
  "itsdangerous",  # For SessionMiddleware
  "google-auth>=2.30.0",
  "google-auth-oauthlib",
  "google-api-python-client",
  "google-cloud-bigquery>=3.25.0",
  "pdfminer.six",
  "beautifulsoup4",
  "bleach",
  "python-dateutil",
  "python-slugify",
  "starlette-exporter",
  "prometheus-client",
  "python-json-logger",
  "redis>=5.0.0",
  # Phase 2: Email categorization & ML
  "scikit-learn>=1.3.0",
  # Phase 5.3: Active learning scheduling
  "APScheduler>=3.10.0",
  # Security analysis
  "idna>=3.4",
  "joblib>=1.3.0",
  "numpy>=1.24.0",
  "scipy>=1.11.0",
  "PyYAML>=6.0",
  "tldextract",
]

[project.optional-dependencies]
test = [
  "pytest>=7.4",
  "pytest-asyncio>=0.21",
  "pytest-cov>=4.1",
  "freezegun>=1.2",
  "requests>=2.31",
]

tracing = [
  "opentelemetry-distro>=0.41b0",
  "opentelemetry-exporter-otlp>=1.20.0",
  "opentelemetry-instrumentation-fastapi>=0.41b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
  "opentelemetry-instrumentation-requests>=0.41b0",
]

[tool.uvicorn]
factory = true

[tool.setuptools]
py-modules = []

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
